feat(replays): adds play/pause event tracking
